update: Refactor to load profile images dynamically
Changed the way the profile image is loaded on the frontend to use the endpoint /image/v1/get-profile-image/{uuid}.
Updated the Contacts component to display the contact's profile image using their UUID.
Modified the LoopContainer component to load the authenticated user's profile image when creating a new "loop".
Added a profile check in the LoopContainer to ensure the profile image is only loaded when the profile is available.
Created the route /get-profile-image/{profileUUID} in the backend to provide the profile image based on the user's UUID.
Refactored the ProfileImageService to ensure proper handling of profile images on the server.